+++ src/V1/Callers/Filters/FilterCodeCaller.php
@@ -52,30 +52,17 @@ use GanbaroDigital\ExceptionHelpers\V1\Callers\Values\CodeCaller;
 class FilterCodeCaller
 {
     /**
-     * recommended list of namespaces to filter from
-     *
-     * DO NOT CHANGE THIS ARRAY FROM YOUR CODE
-     *
-     * @var array
-     */
-    public static $DEFAULT_PARTIALS = [
-        'Checks' => 'Checks',
-        'Exceptions' => 'Exceptions',
-        'Requirements' => 'Requirements',
-    ];
-
-    /**
      * work out who has called a piece of code
      *
      * @param  array $backtrace
      *         the debug_backtrace() return value
-     * @param  array $partialsToFilterOut
-     *         a list of partial namespaces to skip over
+     * @param  array $filterList
+     *         a list of namespaces and classes to skip over
      * @return CodeCaller
      */
-    public function __invoke($backtrace, $partialsToFilterOut = [])
+    public function __invoke($backtrace, $filterList = [])
     {
-        return self::from($backtrace, $partialsToFilterOut);
+        return self::from($backtrace, $filterList);
     }
 
     /**
@@ -83,14 +70,14 @@ class FilterCodeCaller
      *
      * @param  array $backtrace
      *         the debug_backtrace() return value
-     * @param  array $partialsToFilterOut
-     *         a list of partial namespaces to skip over
+     * @param  array $filterList
+     *         a list of namespaces and classes to skip over
      * @return CodeCaller
      */
-    public static function from($backtrace, $partialsToFilterOut = [])
+    public static function from($backtrace, $filterList = [])
     {
         // get the call stack frame that we want to return
-        $frame = FilterBacktrace::from($backtrace, $partialsToFilterOut);
+        $frame = FilterBacktrace::from($backtrace, $filterList);
 
         // convert the stack frame into something that's easier to use
         $retval = new CodeCaller($frame['class'], $frame['function'], $frame['type'], $frame['file'], $frame['line']);